requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
python-dateutil>=2.8.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
